requirements.txt
fastapi
uvicorn
pdfplumber
python-multipart
python-dotenv
pydantic
google-generativeai
